+++ changelog.md
@@ -20,6 +20,7 @@ and this project tries to adhere to [Semantic Versioning](http://semver.org/).
 - Renamed voyState to recMode
 - Fix time in email subject
 - No longer set recMode to off for OTA update
+- Added getSDcardStatus checks
 
 
 ## [4.0.0] - 2023-07-17